Move base members into child
Make the base class have no data members. This is an attempt to make the
child member alignment as predicatable as possible. It is already quite
hacky to cast to an arbitrary child template instance and get at the data.

The variable length data (underlying array) must always come
last.